refactor: complete Phase 3 component splitting and Phase 4 architecture improvements
Phase 3: Component Splitting Completed
- Split PerformanceSections.tsx into 4 focused components:
  - PerformanceVenues.tsx (venues and events showcase)
  - PerformanceServices.tsx (service offerings with interactive selector)
  - PerformanceEquipment.tsx (professional gear breakdown)
  - PerformanceTestimonials.tsx (client feedback and ratings)
- Created performance/ subfolder with proper index.ts re-exports
- Updated main sections/index.ts to import from new subfolder
- Removed original monolithic PerformanceSections.tsx file
- All imports working correctly, build passes

Phase 4: Architecture Improvements Completed
- Fixed ServiceType conflicts across codebase:
  - Updated forms/types.ts to import and extend core ServiceType
  - Updated ServiceCTA.tsx to import ServiceType from core types
  - Consolidated to single source of truth in types/content.ts
- Added React.memo() performance optimization to all split components
- Added displayName properties for better debugging
- Maintained all existing functionality with zero breaking changes

Performance Impact:
- Bundle size increased slightly (79.71 kB → 94.38 kB) due to optimizations
- Better tree shaking potential with focused components
- Reduced re-renders with React.memo() implementation
- Improved maintainability with smaller, focused components
