+++ src/components/sections/performance/index.ts
@@ -1,0 +1,11 @@
+// Performance section components
+export { PerformanceVenues } from './PerformanceVenues'
+export { PerformanceServices } from './PerformanceServices'
+export { PerformanceEquipment } from './PerformanceEquipment'
+export { PerformanceTestimonials } from './PerformanceTestimonials'
+
+// Default exports for backward compatibility
+export { default as PerformanceVenuesDefault } from './PerformanceVenues'
+export { default as PerformanceServicesDefault } from './PerformanceServices'
+export { default as PerformanceEquipmentDefault } from './PerformanceEquipment'
+export { default as PerformanceTestimonialsDefault } from './PerformanceTestimonials'
